requirements.txt
streamlit
python-dotenv
openai
agno
firecrawl
